database_server/face_auth/requirements.txt
flask
flask-cors
deepface
numpy
Pillow
requests
opencv-python